This is an interface function to call and run mcem.

Copyright (C) 2021 The ORF Project Team

Usage

mcem(
  object,
  studentid = "",
  passageid = "",
  nwords.p = "",
  wrc = "",
  time = "",
  k.in = 5,
  reps.in = 2,
  ests.in,
  est = "mcem",
  se = "analytic",
  verbose = FALSE
)

Arguments

object

- reading data object comes from prep function

k.in

- number of passages, default is 5

reps.in

- repeats, default is 2

ests.in

- if not given, mom function will be called and get est.in output

est

- estimator keyword, mcem or mcmc

se

- standard error keyword, default is analytic

verbose

- boolean, if shows the summary, default is FALSE

Value

MCEM list, MCMC list

Details

This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 3 of the License, or (at your option) any later version. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. A copy of the GNU General Public License is available at http://www.gnu.org/licenses/